refactor: #390 `RecommendedReview` business logic을 CQRS pattern 적용하여 분리
